// web/src/types.ts
export interface CapabilityInfo {
  urn: string;
  description?: string;
  online_combs: number;
  tags?: string[];
}